56 update frontend to svelte 5 (#64) (BIG FIX)
* Set update to newest (#55)

* Make ModView compatible with Svelte 5

* Implement (kind of) appropriate way for cache handling

It's still broken tho. Images won't show. The timestamp is in unix time.
Animations are broken now. Fix all of this in the future.

* image loading works now

* Disable image popup for now. Handle later.

* Fix animation

* Fixed issue with mods loading

* Fixed Errors of UninstallDialog.svelte but it's not displaying

The uninstall dialog doesn't appear when pressed on the delete button on
Steamodded or Talisman. Nothing happens, when the delete button is
pressed. Fix later.

* Steamodded and Talisman Dialog still doesn't work tho

* Required Popup & the other mechanisms fixed now

Finally..

* Clean up code

* Clean up more code

* Update deno dependencies

* Remove console logs

* Update cargo dependencies

* Update Cargo.lock

* Fixed issue with git

* Update Markdown handling & Style